requirements.txt
pytorch
numpy
pandas
scikit-learn==0.23.2